app_components/ds/requirements.txt
pandas==2.2.3
numpy== 2.0.2
matplotlib== 3.9.2
scikit-learn== 1.5.2
SQLAlchemy==2.0.36
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.13.2
xgboost==2.1.2
load-dotenv==0.1.0
psycopg2==2.9.10
